+++ modal/functions.py
@@ -35,7 +35,7 @@ from ._output import OutputManager
 from ._pty import get_pty_info
 from ._resolver import Resolver
 from ._resources import convert_fn_config_to_resources_config
-from ._serialization import serialize
+from ._serialization import serialize, serialize_proto_params
 from ._utils.async_utils import (
     TaskContext,
     synchronize_api,
@@ -52,7 +52,7 @@ from ._utils.function_utils import (
     is_async,
 )
 from ._utils.grpc_utils import retry_transient_errors
-from ._utils.mount_utils import validate_mount_points, validate_volumes
+from ._utils.mount_utils import validate_network_file_systems, validate_volumes
 from .call_graph import InputInfo, _reconstruct_call_graph
 from .client import _Client
 from .cloud_bucket_mount import _CloudBucketMount, cloud_bucket_mounts_to_proto
@@ -272,6 +272,7 @@ class _FunctionSpec:
     memory: Optional[Union[int, Tuple[int, int]]]
     ephemeral_disk: Optional[int]
     scheduler_placement: Optional[SchedulerPlacement]
+    _experimental_gpus: Sequence[GPU_T]
 
 
 class _Function(_Object, type_prefix="fu"):
@@ -305,6 +306,8 @@ class _Function(_Object, type_prefix="fu"):
     # TODO (elias): remove _parent. In case of instance functions, and methods bound on those,
     #  this references the parent class-function and is used to infer the client for lazy-loaded methods
     _parent: Optional["_Function"] = None
+
+    _class_parameter_info: Optional["api_pb2.ClassParameterInfo"] = None
 
     def _bind_method(
         self,
@@ -385,7 +388,7 @@ class _Function(_Object, type_prefix="fu"):
         fun._tag = full_name
         fun._raw_f = partial_function.raw_f
         fun._info = FunctionInfo(
-            partial_function.raw_f, cls=user_cls, serialized=class_service_function.info.is_serialized()
+            partial_function.raw_f, user_cls=user_cls, serialized=class_service_function.info.is_serialized()
         )  # needed for .local()
         fun._use_method_name = method_name
         fun._app = class_service_function._app
@@ -495,6 +498,7 @@ class _Function(_Object, type_prefix="fu"):
         block_network: bool = False,
         max_inputs: Optional[int] = None,
         ephemeral_disk: Optional[int] = None,
+        _experimental_gpus: Sequence[GPU_T] = [],
     ) -> None:
         """mdmd:hidden"""
         tag = info.get_tag()
@@ -508,7 +512,7 @@ class _Function(_Object, type_prefix="fu"):
                 )
         else:
             # must be a "class service function"
-            assert info.cls
+            assert info.user_cls
             assert not webhook_config
             assert not schedule
 
@@ -566,8 +570,6 @@ class _Function(_Object, type_prefix="fu"):
                 "Retries for generator functions are deprecated and will soon be removed.",
             )
 
-        gpu_config = parse_gpu_config(gpu)
-
         if proxy:
             # HACK: remove this once we stop using ssh tunnels for this.
             if image:
@@ -585,15 +587,16 @@ class _Function(_Object, type_prefix="fu"):
             memory=memory,
             ephemeral_disk=ephemeral_disk,
             scheduler_placement=scheduler_placement,
+            _experimental_gpus=_experimental_gpus,
         )
 
-        if info.cls and not is_auto_snapshot:
+        if info.user_cls and not is_auto_snapshot:
             # Needed to avoid circular imports
             from .partial_function import _find_callables_for_cls, _PartialFunctionFlags
 
-            build_functions = list(_find_callables_for_cls(info.cls, _PartialFunctionFlags.BUILD).values())
+            build_functions = list(_find_callables_for_cls(info.user_cls, _PartialFunctionFlags.BUILD).values())
             for build_function in build_functions:
-                snapshot_info = FunctionInfo(build_function, cls=info.cls)
+                snapshot_info = FunctionInfo(build_function, user_cls=info.user_cls)
                 snapshot_function = _Function.from_args(
                     snapshot_info,
                     app=None,
@@ -610,6 +613,7 @@ class _Function(_Object, type_prefix="fu"):
                     is_builder_function=True,
                     is_auto_snapshot=True,
                     scheduler_placement=scheduler_placement,
+                    _experimental_gpus=_experimental_gpus,
                 )
                 image = _Image._from_args(
                     base_images={"base": image},
@@ -651,9 +655,7 @@ class _Function(_Object, type_prefix="fu"):
         validated_volumes = [(k, v) for k, v in validated_volumes if isinstance(v, _Volume)]
 
         # Validate NFS
-        if not isinstance(network_file_systems, dict):
-            raise InvalidError("network_file_systems must be a dict[str, NetworkFileSystem] where the keys are paths")
-        validated_network_file_systems = validate_mount_points("Network file system", network_file_systems)
+        validated_network_file_systems = validate_network_file_systems(network_file_systems)
 
         # Validate image
         if image is not None and not isinstance(image, _Image):
@@ -724,7 +726,7 @@ class _Function(_Object, type_prefix="fu"):
                     # serialize at _load time, not function decoration time
                     # otherwise we can't capture a surrounding class for lifetime methods etc.
                     function_serialized = info.serialized_function()
-                    class_serialized = serialize(info.cls) if info.cls is not None else None
+                    class_serialized = serialize(info.user_cls) if info.user_cls is not None else None
                     # Ensure that large data in global variables does not blow up the gRPC payload,
                     # which has maximum size 100 MiB. We set the limit lower for performance reasons.
                     if len(function_serialized) > 16 << 20:  # 16 MiB
@@ -801,7 +803,7 @@ class _Function(_Object, type_prefix="fu"):
                     allow_concurrent_inputs=allow_concurrent_inputs or 0,
                     worker_id=config.get("worker_id"),
                     is_auto_snapshot=is_auto_snapshot,
-                    is_method=bool(info.cls) and not info.is_service_class(),
+                    is_method=bool(info.user_cls) and not info.is_service_class(),
                     checkpointing_enabled=enable_memory_snapshot,
                     is_checkpointing_function=False,
                     object_dependencies=object_dependencies,
@@ -811,6 +813,13 @@ class _Function(_Object, type_prefix="fu"):
                     _experimental_boost=_experimental_boost,
                     scheduler_placement=scheduler_placement.proto if scheduler_placement else None,
                     is_class=info.is_service_class(),
+                    class_parameter_info=info.class_parameter_info(),
+                    _experimental_resources=[
+                        convert_fn_config_to_resources_config(
+                            cpu=cpu, memory=memory, gpu=_experimental_gpu, ephemeral_disk=ephemeral_disk
+                        )
+                        for _experimental_gpu in _experimental_gpus
+                    ],
                 )
                 assert resolver.app_id
                 request = api_pb2.FunctionCreateRequest(
@@ -853,10 +862,14 @@ class _Function(_Object, type_prefix="fu"):
         # hash. We can't use the cloudpickle hash because it's not very stable.
         obj._build_args = dict(  # See get_build_def
             secrets=repr(secrets),
-            gpu_config=repr(gpu_config),
+            gpu_config=repr(parse_gpu_config(gpu)),
             mounts=repr(mounts),
             network_file_systems=repr(network_file_systems),
         )
+        if _experimental_gpus:
+            obj._build_args["experimental_gpus"] = repr(
+                [parse_gpu_config(_experimental_gpu) for _experimental_gpu in _experimental_gpus]
+            )
 
         return obj
 
@@ -890,7 +903,17 @@ class _Function(_Object, type_prefix="fu"):
                     f"The {identity} has not been hydrated with the metadata it needs to run on Modal{reason}."
                 )
             assert self._parent._client.stub
-            serialized_params = serialize((args, kwargs))
+            if (
+                self._parent._class_parameter_info
+                and self._parent._class_parameter_info.format
+                == api_pb2.ClassParameterInfo.PARAM_SERIALIZATION_FORMAT_PROTO
+            ):
+                if args:
+                    # TODO(elias) - We could potentially support positional args as well, if we want to?
+                    raise InvalidError("Can't use positional arguments with strict parameter classes")
+                serialized_params = serialize_proto_params(kwargs, self._parent._class_parameter_info.schema)
+            else:
+                serialized_params = serialize((args, kwargs))
             environment_name = _get_environment_name(None, resolver)
             assert self._parent is not None
             req = api_pb2.FunctionBindParamsRequest(
@@ -1070,6 +1093,7 @@ class _Function(_Object, type_prefix="fu"):
         self._is_method = metadata.is_method
         self._use_function_id = metadata.use_function_id
         self._use_method_name = metadata.use_method_name
+        self._class_parameter_info = metadata.class_parameter_info
 
     def _invocation_function_id(self) -> str:
         return self._use_function_id or self.object_id
@@ -1088,6 +1112,7 @@ class _Function(_Object, type_prefix="fu"):
             use_method_name=self._use_method_name,
             use_function_id=self._use_function_id,
             is_method=self._is_method,
+            class_parameter_info=self._class_parameter_info,
         )
 
     def _set_mute_cancellation(self, value: bool = True):