+++ modal/image.py
@@ -156,11 +156,29 @@ def _flatten_str_args(function_name: str, arg_name: str, args: Tuple[Union[str, 
     return ret
 
 
+def _validate_packages(packages: List[str]) -> bool:
+    """Validates that a list of packages does not contain any command-line options."""
+    return not any(pkg.startswith("-") for pkg in packages)
+
+
+def _warn_invalid_packages(old_command: str) -> None:
+    deprecation_warning(
+        (2024, 7, 3),
+        "Passing flags to `pip` via the `packages` argument of `pip_install` is deprecated."
+        " Please pass flags via the `extra_options` argument instead."
+        "\nNote that this will cause a rebuild of this image layer."
+        " To avoid rebuilding, you can pass the following to `run_commands` instead:"
+        f'\n`image.run_commands("{old_command}")`',
+        show_source=False,
+    )
+
+
 def _make_pip_install_args(
     find_links: Optional[str] = None,  # Passes -f (--find-links) pip install
     index_url: Optional[str] = None,  # Passes -i (--index-url) to pip install
     extra_index_url: Optional[str] = None,  # Passes --extra-index-url to pip install
     pre: bool = False,  # Passes --pre (allow pre-releases) to pip install
+    extra_options: str = "",  # Additional options to pass to pip install, e.g. "--no-build-isolation --no-clean"
 ) -> str:
     flags = [
         ("--find-links", find_links),  # TODO(erikbern): allow multiple?
@@ -170,7 +188,12 @@ def _make_pip_install_args(
 
     args = " ".join(f"{flag} {shlex.quote(value)}" for flag, value in flags if value is not None)
     if pre:
-        args += " --pre"
+        args += " --pre"  # TODO: remove extra whitespace in future image builder version
+
+    if extra_options:
+        if args:
+            args += " "
+        args += f"{extra_options}"
 
     return args
 
@@ -322,6 +345,8 @@ class _Image(_Object, type_prefix="im"):
                 build_function_id = build_function.object_id
 
                 globals = build_function._get_info().get_globals()
+                attrs = build_function._get_info().get_cls_var_attrs()
+                globals = {**globals, **attrs}
                 filtered_globals = {}
                 for k, v in globals.items():
                     if isfunction(v):
@@ -517,17 +542,36 @@ class _Image(_Object, type_prefix="im"):
         index_url: Optional[str] = None,  # Passes -i (--index-url) to pip install
         extra_index_url: Optional[str] = None,  # Passes --extra-index-url to pip install
         pre: bool = False,  # Passes --pre (allow pre-releases) to pip install
+        extra_options: str = "",  # Additional options to pass to pip install, e.g. "--no-build-isolation --no-clean"
         force_build: bool = False,  # Ignore cached builds, similar to 'docker build --no-cache'
         secrets: Sequence[_Secret] = [],
         gpu: GPU_T = None,
     ) -> "_Image":
         """Install a list of Python packages using pip.
 
-        **Example**
+        **Examples**
 
+        Simple installation:
         ```python
         image = modal.Image.debian_slim().pip_install("click", "httpx~=0.23.3")
         ```
+
+        More complex installation:
+        ```python
+        image = (
+            modal.Image.from_registry(
+                "nvidia/cuda:12.2.0-devel-ubuntu22.04", add_python="3.11"
+            )
+            .pip_install(
+                "ninja",
+                "packaging",
+                "wheel",
+                "transformers==4.40.2",
+            )
+            .pip_install(
+                "flash-attn==2.5.8", extra_options="--no-build-isolation"
+            )
+        )
         """
         pkgs = _flatten_str_args("pip_install", "packages", packages)
         if not pkgs:
@@ -535,8 +579,10 @@ class _Image(_Object, type_prefix="im"):
 
         def build_dockerfile(version: ImageBuilderVersion) -> DockerfileSpec:
             package_args = " ".join(shlex.quote(pkg) for pkg in sorted(pkgs))
-            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre)
+            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre, extra_options)
             commands = ["FROM base", f"RUN python -m pip install {package_args} {extra_args}"]
+            if not _validate_packages(pkgs):
+                _warn_invalid_packages(commands[-1].split("RUN ")[-1])
             if version > "2023.12":  # Back-compat for legacy trailing space with empty extra_args
                 commands = [cmd.strip() for cmd in commands]
             return DockerfileSpec(commands=commands, context_files={})
@@ -558,6 +604,7 @@ class _Image(_Object, type_prefix="im"):
         index_url: Optional[str] = None,  # Passes -i (--index-url) to pip install
         extra_index_url: Optional[str] = None,  # Passes --extra-index-url to pip install
         pre: bool = False,  # Passes --pre (allow pre-releases) to pip install
+        extra_options: str = "",  # Additional options to pass to pip install, e.g. "--no-build-isolation --no-clean"
         gpu: GPU_T = None,
         secrets: Sequence[_Secret] = [],
         force_build: bool = False,  # Ignore cached builds, similar to 'docker build --no-cache'
@@ -633,7 +680,7 @@ class _Image(_Object, type_prefix="im"):
                     f"(echo 'GITLAB_TOKEN env var not set by provided modal.Secret(s): {secret_names}' && exit 1)\"",
                 )
 
-            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre)
+            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre, extra_options)
             commands.extend(["RUN apt-get update && apt-get install -y git"])
             commands.extend([f'RUN python3 -m pip install "{url}" {extra_args}' for url in install_urls])
             if version > "2023.12":  # Back-compat for legacy trailing space with empty extra_args
@@ -658,6 +705,7 @@ class _Image(_Object, type_prefix="im"):
         index_url: Optional[str] = None,  # Passes -i (--index-url) to pip install
         extra_index_url: Optional[str] = None,  # Passes --extra-index-url to pip install
         pre: bool = False,  # Passes --pre (allow pre-releases) to pip install
+        extra_options: str = "",  # Additional options to pass to pip install, e.g. "--no-build-isolation --no-clean"
         force_build: bool = False,  # Ignore cached builds, similar to 'docker build --no-cache'
         secrets: Sequence[_Secret] = [],
         gpu: GPU_T = None,
@@ -670,7 +718,7 @@ class _Image(_Object, type_prefix="im"):
 
             null_find_links_arg = " " if version == "2023.12" else ""
             find_links_arg = f" -f {find_links}" if find_links else null_find_links_arg
-            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre)
+            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre, extra_options)
 
             commands = [
                 "FROM base",
@@ -698,6 +746,7 @@ class _Image(_Object, type_prefix="im"):
         index_url: Optional[str] = None,  # Passes -i (--index-url) to pip install
         extra_index_url: Optional[str] = None,  # Passes --extra-index-url to pip install
         pre: bool = False,  # Passes --pre (allow pre-releases) to pip install
+        extra_options: str = "",  # Additional options to pass to pip install, e.g. "--no-build-isolation --no-clean"
         force_build: bool = False,  # Ignore cached builds, similar to 'docker build --no-cache'
         secrets: Sequence[_Secret] = [],
         gpu: GPU_T = None,
@@ -734,7 +783,7 @@ class _Image(_Object, type_prefix="im"):
                     if dep_group_name in optionals:
                         dependencies.extend(optionals[dep_group_name])
 
-            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre)
+            extra_args = _make_pip_install_args(find_links, index_url, extra_index_url, pre, extra_options)
             package_args = " ".join(shlex.quote(pkg) for pkg in sorted(dependencies))
             commands = ["FROM base", f"RUN python -m pip install {package_args} {extra_args}"]
             if version > "2023.12":  # Back-compat for legacy trailing space
@@ -853,7 +902,7 @@ class _Image(_Object, type_prefix="im"):
             base_images={"base": self},
             dockerfile_function=build_dockerfile,
             secrets=secrets,
-            gpu_config=parse_gpu_config(gpu, raise_on_true=False),
+            gpu_config=parse_gpu_config(gpu),
             context_mount=context_mount,
             force_build=self.force_build or force_build,
         )
@@ -899,7 +948,7 @@ class _Image(_Object, type_prefix="im"):
             base_images={"base": self},
             dockerfile_function=build_dockerfile,
             secrets=secrets,
-            gpu_config=parse_gpu_config(gpu, raise_on_true=False),
+            gpu_config=parse_gpu_config(gpu),
             force_build=self.force_build or force_build,
         )