+++ test/app_test.py
@@ -6,7 +6,7 @@ import pytest
 from google.protobuf.empty_pb2 import Empty
 from grpclib import GRPCError, Status
 
-from modal import App, Dict, Image, Mount, Queue, Secret, Stub, Volume, web_endpoint
+from modal import App, Dict, Image, Mount, Secret, Stub, Volume, web_endpoint
 from modal.app import list_apps  # type: ignore
 from modal.exception import DeprecationError, ExecutionError, InvalidError, NotFoundError
 from modal.partial_function import _parse_custom_domains
@@ -14,15 +14,6 @@ from modal.runner import deploy_app, deploy_stub
 from modal_proto import api_pb2
 
 from .supports import module_1, module_2
-
-
-@pytest.mark.asyncio
-async def test_kwargs(servicer, client):
-    with pytest.raises(DeprecationError):
-        App(
-            d=Dict.new(),
-            q=Queue.new(),
-        )
 
 
 @pytest.mark.asyncio